services/cart/requirements/base.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.0.10
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-environ==0.8.1
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.1
sqlparse==0.4.2
whitenoise==6.1.0
wrapt==1.14.1
